+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.0.0-RC2 - 2022-05-03
+
+## Added
+- Provided a migration to update the namespace from percipioglobal to percipiolondon
+
+## 4.0.0-RC1 - 2022-05-02
+
+## Changed
+- Create release candidate of 4.0.0-beta.2
+
+## 4.0.0-beta.2 - 2022-05-02
+
+### Added
+- Updated PHPDocs
+
+### Fixed
+- Fixed the editableTable config
+
 ## 4.0.0-beta.1 - 2022-04-20
 
 ### Added